fix(ci): python-coverage-shards — split auth_mounts_claude + overlay test parts under 1500-line maintainability limit
The first-party line-limit gate (test_first_party_code_files_stay_under_line_limit)
failed: review-comment fixes grew auth_mounts_claude.py (1562), and the overlay
test parts part_001 (1583) and part_002 (1717) past the 1500-line cap.

- Extract the generic /proc/mounts lowerdir-discovery helpers
  (_unescape_proc_mount_field, _overlay_lowerdir_from_proc_mounts,
  iter_overlay_lowerdirs) + _PROC_MOUNTS into new awf.node.auth_mounts_overlay;
  auth_mounts_claude re-exports them so the stable import surface is unchanged.
- Split the oversized overlay test files into part_003 (proc-mount/reconcile
  tests from part_002) and part_004 (retry/EBUSY tests from part_001), keeping
  the shared FakeOverlayMounter/_seed_host_claude helpers in part_001.
